Se realizan los cambios solicitados
