+++ src/firebase/client.tsx
@@ -352,44 +352,3 @@ export function updateProfileCover(image: File) {
   const task = storageRef.put(image);
   return task;
 }
-export async function AddNewCallClub(
-  clubId: string,
-  ageRequired: string,
-  posRequired: string,
-  startDate: string,
-  endDate: string,
-  extraDetails: string
-) {
-  try {
-    await firestore.collection(COLLECTIONS.CALLS).add({
-      clubId,
-      ageRequired,
-      posRequired,
-      startDate: firebase.firestore.Timestamp.fromDate(new Date(startDate)),
-      endDate: firebase.firestore.Timestamp.fromDate(new Date(endDate)),
-      extraDetails,
-      postulatedPlayers: {},
-    });
-    return true;
-  } catch (e) {
-    return false;
-  }
-}
-export const getMyCallsData = (
-  callback: React.Dispatch<
-    React.SetStateAction<firebase.firestore.DocumentData | undefined>
-  >
-) => {
-  let id = auth.currentUser?.uid;
-  return firestore
-    .collection(COLLECTIONS.CALLS)
-    .where("clubId", "==", id)
-    .orderBy("endDate", "desc")
-    .onSnapshot((snapshot) => {
-      const newData = snapshot.docs.map((doc) => {
-        const data = { id: doc.id, ...doc.data() };
-        return data;
-      });
-      callback(newData);
-    });
-};
