+++ src/pages/CallsListClub/styles.module.css
@@ -42,3 +42,14 @@
   border-bottom-color: black !important;
   color: #315c79 !important;
 }
+
+.message_not_found {
+  font-size: 12px;
+  --ion-background-color: #d0dde1;
+  color: #444;
+  opacity: 80%;
+}
+.container_message {
+  text-align: center;
+  padding-top: 20px;
+}
