+++ src/utils/ConvertFunctions.tsx
@@ -1,0 +1,8 @@
+import firebase from "firebase/app";
+export function converterDate(date: firebase.firestore.Timestamp) {
+  let convertDate = date.toDate();
+  let newDate = `${convertDate.getDate()} / ${
+    convertDate.getUTCMonth() + 1
+  } / ${convertDate.getUTCFullYear()}`;
+  return newDate;
+}
